requirements.txt
flask 
pillow 
numpy 
matplotlib
